Optimized HAB Mini App Detail
