+++ lib/MiniApps/HotAirBalloons/Screens/DetailScreen/widgets/HABDetailScreenFlightDetailsTab.dart
@@ -1,16 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_vector_icons/flutter_vector_icons.dart';
 
+import 'package:flutter_uis/configs/AppDimensions.dart';
+
 import 'package:flutter_uis/UI.dart';
 
 import '../../../configs/theme.dart' as theme;
 import '../../../data/flights.dart' as data;
 
-import '../Dimensions.dart';
-import 'RowInfo.dart';
+import 'HABDetailScreenRowInfo.dart';
 
-class FlightDetailsTab extends StatelessWidget {
-  FlightDetailsTab(this.flight);
+class HABDetailScreenFlightDetailsTab extends StatelessWidget {
+  HABDetailScreenFlightDetailsTab(this.flight);
   final data.HABFlight flight;
 
   @override
@@ -18,13 +19,13 @@ class FlightDetailsTab extends StatelessWidget {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: <Widget>[
-        RowInfo("Basket Size", flight.basketSize, Feather.users),
-        RowInfo("Flight Time", flight.time, Feather.clock),
-        RowInfo("Price", flight.price, Feather.tag),
+        HABDetailScreenRowInfo("Basket Size", flight.basketSize, Feather.users),
+        HABDetailScreenRowInfo("Flight Time", flight.time, Feather.clock),
+        HABDetailScreenRowInfo("Price", flight.price, Feather.tag),
         Padding(
           padding: EdgeInsets.symmetric(
-            horizontal: Dimensions.padding * 5,
-            vertical: Dimensions.padding * 4,
+            horizontal: AppDimensions.padding * 5,
+            vertical: AppDimensions.padding * 4,
           ),
           child: Text(
             flight.desc,
@@ -39,8 +40,8 @@ class FlightDetailsTab extends StatelessWidget {
           width: double.infinity,
           constraints: BoxConstraints(maxWidth: 500),
           margin: EdgeInsets.symmetric(
-            vertical: Dimensions.padding * 2,
-            horizontal: Dimensions.padding * 3,
+            vertical: AppDimensions.padding * 2,
+            horizontal: AppDimensions.padding * 3,
           ),
           decoration: BoxDecoration(
             borderRadius: BorderRadius.circular(12.0),
@@ -59,7 +60,7 @@ class FlightDetailsTab extends StatelessWidget {
             ),
             onPressed: () {},
             color: theme.primary,
-            padding: EdgeInsets.symmetric(vertical: Dimensions.padding * 2),
+            padding: EdgeInsets.symmetric(vertical: AppDimensions.padding * 2),
             textColor: Colors.white,
             child: Text(
               "BOOK NOW",
