+++ lib/MiniApps/HotAirBalloons/Screens/DetailScreen/widgets/HABDetailScreenPostFlightInfoTab.dart
@@ -2,29 +2,29 @@ import 'package:flutter/material.dart';
 import 'package:flutter_vector_icons/flutter_vector_icons.dart';
 
 import '../../../data/flights.dart' as data;
-import 'RowInfo.dart';
+import 'HABDetailScreenRowInfo.dart';
 
-class PostFlightInfoTab extends StatelessWidget {
-  const PostFlightInfoTab(this.flight);
+class HABDetailScreenPostFlightInfoTab extends StatelessWidget {
+  const HABDetailScreenPostFlightInfoTab(this.flight);
   final data.HABFlight flight;
 
   @override
   Widget build(BuildContext context) {
     return Column(
       children: [
-        RowInfo(
+        HABDetailScreenRowInfo(
           "Celebration",
           flight.celebration,
           Entypo.drink,
           fullWidth: true,
         ),
-        RowInfo(
+        HABDetailScreenRowInfo(
           "Flight Certificate",
           flight.certificate,
           MaterialCommunityIcons.certificate,
           fullWidth: true,
         ),
-        RowInfo(
+        HABDetailScreenRowInfo(
           "Service back to your hotel",
           flight.serviceBack,
           MaterialCommunityIcons.bus_side,
