+++ lib/MiniApps/HotAirBalloons/Screens/DetailScreen/widgets/HABDetailScreenPreFlightInfoTab.dart
@@ -2,29 +2,29 @@ import 'package:flutter/material.dart';
 import 'package:flutter_vector_icons/flutter_vector_icons.dart';
 
 import '../../../data/flights.dart' as data;
-import 'RowInfo.dart';
+import 'HABDetailScreenRowInfo.dart';
 
-class PreFlightInfoTab extends StatelessWidget {
-  const PreFlightInfoTab(this.flight);
+class HABDetailScreenPreFlightInfoTab extends StatelessWidget {
+  const HABDetailScreenPreFlightInfoTab(this.flight);
   final data.HABFlight flight;
 
   @override
   Widget build(BuildContext context) {
     return Column(
       children: [
-        RowInfo(
+        HABDetailScreenRowInfo(
           "Pick-up from your hotel",
           flight.pickUp,
           MaterialCommunityIcons.bus_side,
           fullWidth: true,
         ),
-        RowInfo(
+        HABDetailScreenRowInfo(
           flight.breakFastType,
           flight.breakFast,
           Feather.coffee,
           fullWidth: true,
         ),
-        RowInfo(
+        HABDetailScreenRowInfo(
           "Watching the inflation",
           flight.inflation,
           MaterialCommunityIcons.airballoon,
